Use `Nut` in first example
When first trying to understand frameworks, I first skim code examples
because that tells me a lot more and are clearer than the text. We,
fans of the ML family, often look at the type signatures as a form of
documentation. For me, adding the type of the component makes it clearer
what’s going on—even if it is explained in bullet #2.